Convert target and unit tests to arc.
The only real item of note here is that the parserDelegate ivar declaration has been removed. This is to avoid using the ownership qualification __unsafe_unretained that ARC requires for assign properties that have an ivar defined. Omitting the ivar is, IMO, a better approach than qualifying the ivar since this ensures that ownership semantics are in one place -- on the properly declaration.